requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
flower==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanize==4.9.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
kombu==5.3.4
mongoengine==0.27.0
mongomock==4.1.2
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pymongo==4.6.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
ruff==0.1.9
sendgrid==6.11.0
sentinels==1.0.0
six==1.16.0
sniffio==1.3.0
starkbank-ecdsa==2.2.0
starlette==0.32.0.post1
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.25.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
